+++ src/pluto_main/Table_Country.cpp
@@ -204,7 +204,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)30,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(30,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -218,7 +218,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Define.c_str(), (unsigned long) min((size_t)30,m_Define.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Define.c_str(), (unsigned long) min(30,m_Define.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -245,7 +245,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[11];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_country_code.c_str(), (unsigned long) min((size_t)5,m_country_code.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_country_code.c_str(), (unsigned long) min(5,m_country_code.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
