+++ src/pluto_main/Table_DeviceTemplate.cpp
@@ -620,7 +620,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)30,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(30,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -634,7 +634,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[5000000];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comments.c_str(), (unsigned long) min((size_t)16777215,m_Comments.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comments.c_str(), (unsigned long) min(16777215,m_Comments.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -674,7 +674,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Define.c_str(), (unsigned long) min((size_t)30,m_Define.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Define.c_str(), (unsigned long) min(30,m_Define.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -714,7 +714,7 @@ if (is_null[8])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_CommandLine.c_str(), (unsigned long) min((size_t)100,m_CommandLine.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_CommandLine.c_str(), (unsigned long) min(100,m_CommandLine.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -806,7 +806,7 @@ if (is_null[15])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_DestinationDir.c_str(), (unsigned long) min((size_t)50,m_DestinationDir.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_DestinationDir.c_str(), (unsigned long) min(50,m_DestinationDir.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -872,7 +872,7 @@ if (is_null[20])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ConfigureScript.c_str(), (unsigned long) min((size_t)100,m_ConfigureScript.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ConfigureScript.c_str(), (unsigned long) min(100,m_ConfigureScript.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -886,7 +886,7 @@ if (is_null[21])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_ManufacturerURL.c_str(), (unsigned long) min((size_t)100,m_ManufacturerURL.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_ManufacturerURL.c_str(), (unsigned long) min(100,m_ManufacturerURL.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -900,7 +900,7 @@ if (is_null[22])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_InternalURLSuffix.c_str(), (unsigned long) min((size_t)50,m_InternalURLSuffix.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_InternalURLSuffix.c_str(), (unsigned long) min(50,m_InternalURLSuffix.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -914,7 +914,7 @@ if (is_null[23])
 return "NULL";
 
 char *buf = new char[101];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_WizardURL.c_str(), (unsigned long) min((size_t)50,m_WizardURL.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_WizardURL.c_str(), (unsigned long) min(50,m_WizardURL.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -967,7 +967,7 @@ if (is_null[27])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL.c_str(), (unsigned long) min((size_t)255,m_HelpURL.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL.c_str(), (unsigned long) min(255,m_HelpURL.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -1033,7 +1033,7 @@ if (is_null[32])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
