+++ src/pluto_main/Table_Package.cpp
@@ -425,7 +425,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min((size_t)255,m_Description.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Description.c_str(), (unsigned long) min(255,m_Description.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -491,7 +491,7 @@ if (is_null[6])
 return "NULL";
 
 char *buf = new char[201];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_HomePage.c_str(), (unsigned long) min((size_t)100,m_HomePage.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_HomePage.c_str(), (unsigned long) min(100,m_HomePage.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -531,7 +531,7 @@ if (is_null[9])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL.c_str(), (unsigned long) min((size_t)255,m_HelpURL.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL.c_str(), (unsigned long) min(255,m_HelpURL.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -545,7 +545,7 @@ if (is_null[10])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL_UsersManual.c_str(), (unsigned long) min((size_t)255,m_HelpURL_UsersManual.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL_UsersManual.c_str(), (unsigned long) min(255,m_HelpURL_UsersManual.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -559,7 +559,7 @@ if (is_null[11])
 return "NULL";
 
 char *buf = new char[511];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL_ProgrammersGuide.c_str(), (unsigned long) min((size_t)255,m_HelpURL_ProgrammersGuide.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_HelpURL_ProgrammersGuide.c_str(), (unsigned long) min(255,m_HelpURL_ProgrammersGuide.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -573,7 +573,7 @@ if (is_null[12])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comments.c_str(), (unsigned long) min((size_t)65535,m_Comments.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Comments.c_str(), (unsigned long) min(65535,m_Comments.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -639,7 +639,7 @@ if (is_null[17])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
