+++ src/pluto_main/Table_PostalCode.cpp
@@ -259,7 +259,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[21];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_PostalCode.c_str(), (unsigned long) min((size_t)10,m_PostalCode.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_PostalCode.c_str(), (unsigned long) min(10,m_PostalCode.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -273,7 +273,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[21];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Long.c_str(), (unsigned long) min((size_t)10,m_Long.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Long.c_str(), (unsigned long) min(10,m_Long.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -287,7 +287,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[21];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Lat.c_str(), (unsigned long) min((size_t)10,m_Lat.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Lat.c_str(), (unsigned long) min(10,m_Lat.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -301,7 +301,7 @@ if (is_null[4])
 return "NULL";
 
 char *buf = new char[121];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_City.c_str(), (unsigned long) min((size_t)60,m_City.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_City.c_str(), (unsigned long) min(60,m_City.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -315,7 +315,7 @@ if (is_null[5])
 return "NULL";
 
 char *buf = new char[5];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_State.c_str(), (unsigned long) min((size_t)2,m_State.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_State.c_str(), (unsigned long) min(2,m_State.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -329,7 +329,7 @@ if (is_null[6])
 return "NULL";
 
 char *buf = new char[121];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_County.c_str(), (unsigned long) min((size_t)60,m_County.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_County.c_str(), (unsigned long) min(60,m_County.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
